+++ CHANGELOG.md
@@ -13,11 +13,17 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 - Support Laravel 9
 
+## v5.39.2
+
+### Fixed
+
+- Handle expired subscriber keys in `RedisStorageManager::subscribersByTopic()` https://github.com/nuwave/lighthouse/issues/2035
+
 ## v5.39.1
 
 ### Fixed
 
-- Remove unused configuration option `batched_queries`
+- Remove unused configuration option `batched_queries` https://github.com/nuwave/lighthouse/issues/2046
 
 ## v5.39.0